[docs] Rearranged Summary examples
